[feat] 將 header 右側使用者選單獨立為單獨元件供重複使用，並加上顯示使用者頭像的功能與修正Profile Link的連結
